windsurf-project/requirements.txt
google-adk
langchain
python-dotenv
requests

